Count Number of Digits in Integer using While loop
This Java program counts the number of digits in a given integer. It takes an integer input from the user using the Scanner class. A while loop repeatedly divides the number by 10, incrementing the count variable until the number becomes zero. Finally, the program prints the total count of digits.