Weekday Name from Numbers using Switch Case.
This Java program takes a number (1-7) as input and prints the corresponding weekday name. It uses a switch statement to map each number to a specific day of the week. If the input is outside the range of 1-7, it displays "Invalid Input". Finally, it prints the determined weekday name and closes the Scanner object.