Natural Numbers and Their Sum.
This Java program calculates the sum of the digits of a given number. It takes an integer input from the user and uses a while loop to extract each digit by using the modulo operation (num % 10) and adds it to the sum. The number is then divided by 10 to remove the last digit. Finally, it prints the sum of the digits.